008_fullstack_projects/01_task_app/backend/requirements.txt
coreapi==2.3.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
pip-chill==1.0.3
